+++ mlir/tools/quantum-opt/quantum-opt.cpp
@@ -32,6 +32,7 @@ int main(int argc, char **argv)
     mlir::registerPass(catalyst::createQuantumBufferizationPass);
     mlir::registerPass(catalyst::createQuantumConversionPass);
     mlir::registerPass(catalyst::createEmitCatalystPyInterfacePass);
+    mlir::registerPass(catalyst::createCopyGlobalMemRefPass);
 
     mlir::DialectRegistry registry;
     mlir::registerAllDialects(registry);