+++ runtime/lib/backend/QubitManager.hpp
@@ -18,8 +18,8 @@
 #include <map>
 
 #include "Exception.hpp"
-#include "LightningUtils.hpp"
 #include "Types.h"
+#include "Utils.hpp"
 
 namespace Catalyst::Runtime::Simulator {