+++ runtime/lib/backend/Utils.hpp
@@ -23,8 +23,6 @@
 #include "Types.h"
 
 #if __has_include("StateVectorKokkos.hpp")
-// this macro is used in the C++ test suite
-#define _KOKKOS
 #include "UtilKokkos.hpp"
 #else
 #include "Util.hpp"