+++ runtime/tests/Test_LightningGateSet.cpp
@@ -14,26 +14,20 @@
 
 #include <cmath>
 
-#include "LightningUtils.hpp"
 #include "QuantumDevice.hpp"
 #include "RuntimeCAPI.h"
+#include "Utils.hpp"
 
-#include <catch2/catch.hpp>
-
-#if defined(_KOKKOS)
-#include "LightningKokkosSimulator.hpp"
-#else
-#include "LightningSimulator.hpp"
-#endif
+#include "TestUtils.hpp"
 
 using namespace Pennylane;
 
 using namespace Catalyst::Runtime;
 using namespace Catalyst::Runtime::Simulator;
 
-TEST_CASE("Identity Gate tests", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("Identity Gate tests", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 10;
@@ -63,9 +57,9 @@ TEST_CASE("Identity Gate tests", "[lightning]")
     CHECK(sum == std::complex<double>{0, 0});
 }
 
-TEST_CASE("PauliX Gate tests num_qubits=1", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("PauliX Gate tests num_qubits=1", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 1;
@@ -87,9 +81,9 @@ TEST_CASE("PauliX Gate tests num_qubits=1", "[lightning]")
 }
 
 // 1-qubit operations
-TEST_CASE("PauliX Gate tests num_qubits=3", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("PauliX Gate tests num_qubits=3", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 3;
@@ -122,9 +116,9 @@ TEST_CASE("PauliX Gate tests num_qubits=3", "[lightning]")
     CHECK(sum == std::complex<double>{0, 0});
 }
 
-TEST_CASE("PauliY Gate tests num_qubits=1", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("PauliY Gate tests num_qubits=1", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 1;
@@ -145,9 +139,9 @@ TEST_CASE("PauliY Gate tests num_qubits=1", "[lightning]")
     CHECK(state.at(1) == std::complex<double>{0, 1});
 }
 
-TEST_CASE("PauliY Gate tests num_qubits=2", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("PauliY Gate tests num_qubits=2", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -171,9 +165,9 @@ TEST_CASE("PauliY Gate tests num_qubits=2", "[lightning]")
     CHECK(state.at(3) == std::complex<double>{-1, 0});
 }
 
-TEST_CASE("PauliZ Gate tests num_qubits=2", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("PauliZ Gate tests num_qubits=2", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -197,9 +191,9 @@ TEST_CASE("PauliZ Gate tests num_qubits=2", "[lightning]")
     CHECK(state.at(3) == std::complex<double>{0, 0});
 }
 
-TEST_CASE("Hadamard Gate tests num_qubits=2", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("Hadamard Gate tests num_qubits=2", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -224,9 +218,9 @@ TEST_CASE("Hadamard Gate tests num_qubits=2", "[lightning]")
     CHECK(state.at(3) == state.at(0));
 }
 
-TEST_CASE("Hadamard Gate tests num_qubits=3", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("Hadamard Gate tests num_qubits=3", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -251,9 +245,9 @@ TEST_CASE("Hadamard Gate tests num_qubits=3", "[lightning]")
     CHECK(state.at(3) == state.at(0));
 }
 
-TEST_CASE("MIX Gate test R(X,Y,Z) num_qubits=1,4", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("MIX Gate test R(X,Y,Z) num_qubits=1,4", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubit = n
     constexpr size_t n = 4;
@@ -299,9 +293,9 @@ TEST_CASE("MIX Gate test R(X,Y,Z) num_qubits=1,4", "[lightning]")
     CHECK(state.at(15) == std::complex<double>{0, 0});
 }
 
-TEST_CASE("test PhaseShift num_qubits=2", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("test PhaseShift num_qubits=2", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubit = n
     constexpr size_t n = 2;
@@ -334,9 +328,9 @@ TEST_CASE("test PhaseShift num_qubits=2", "[lightning]")
 }
 
 // 2-qubit operations
-TEST_CASE("CNOT Gate tests num_qubits=2 [0,1]", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("CNOT Gate tests num_qubits=2 [0,1]", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -360,9 +354,9 @@ TEST_CASE("CNOT Gate tests num_qubits=2 [0,1]", "[lightning]")
     CHECK(state.at(3) == std::complex<double>{1, 0});
 }
 
-TEST_CASE("CNOT Gate tests num_qubits=2 [1,0]", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("CNOT Gate tests num_qubits=2 [1,0]", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -386,9 +380,9 @@ TEST_CASE("CNOT Gate tests num_qubits=2 [1,0]", "[lightning]")
     CHECK(state.at(3) == std::complex<double>{0, 0});
 }
 
-TEST_CASE("MIX Gate test CR(X, Y, Z) num_qubits=1,4", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("MIX Gate test CR(X, Y, Z) num_qubits=1,4", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubit = n
     constexpr size_t n = 4;
@@ -431,9 +425,9 @@ TEST_CASE("MIX Gate test CR(X, Y, Z) num_qubits=1,4", "[lightning]")
     CHECK(state.at(15) == std::complex<double>{0, 0});
 }
 
-TEST_CASE("CRot", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("CRot", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -460,9 +454,9 @@ TEST_CASE("CRot", "[lightning]")
     CHECK(state[3].imag() == Approx(-0.4844562109).epsilon(1e-5));
 }
 
-TEST_CASE("CSWAP test", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("CSWAP test", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 3;
@@ -484,9 +478,9 @@ TEST_CASE("CSWAP test", "[lightning]")
     CHECK(state[5].imag() == Approx(0).epsilon(1e-5));
 }
 
-TEST_CASE("IsingXY Gate tests num_qubits=2 [1,0]", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("IsingXY Gate tests num_qubits=2 [1,0]", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -516,9 +510,9 @@ TEST_CASE("IsingXY Gate tests num_qubits=2 [1,0]", "[lightning]")
     CHECK(state[3].imag() == Approx(0).epsilon(1e-5));
 }
 
-TEST_CASE("Toffoli test", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("Toffoli test", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 3;
@@ -549,9 +543,9 @@ TEST_CASE("Toffoli test", "[lightning]")
     CHECK(state[7].imag() == Approx(0).epsilon(1e-5));
 }
 
-TEST_CASE("MultiRZ test", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("MultiRZ test", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -576,9 +570,9 @@ TEST_CASE("MultiRZ test", "[lightning]")
     CHECK(state[2].imag() == Approx(0).epsilon(1e-5));
 }
 
-TEST_CASE("MatrixOperation test with 2-qubit", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("MatrixOperation test with 2-qubit", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 2;
@@ -618,9 +612,9 @@ TEST_CASE("MatrixOperation test with 2-qubit", "[lightning]")
     CHECK(state[3].imag() == Approx(-0.187075).epsilon(1e-5));
 }
 
-TEST_CASE("MatrixOperation test with 3-qubit", "[lightning]")
+TEMPLATE_LIST_TEST_CASE("MatrixOperation test with 3-qubit", "[GateSet]", SimTypes)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubits = n
     constexpr size_t n = 3;
@@ -631,6 +625,7 @@ TEST_CASE("MatrixOperation test with 3-qubit", "[lightning]")
         Qs[i] = sim->AllocateQubit();
     }
 
+    sim->StartTapeRecording();
     sim->NamedOperation("Hadamard", {}, {Qs[0]}, false);
     sim->NamedOperation("Toffoli", {}, {Qs[0], Qs[1], Qs[2]}, false);
 
@@ -646,6 +641,7 @@ TEST_CASE("MatrixOperation test with 3-qubit", "[lightning]")
         {0.327486634260, 0.130699704247}, {0.299805387808, 0.150417378569},
     };
     sim->MatrixOperation(matrix, wires, false);
+    sim->StopTapeRecording();
 
     std::vector<std::complex<double>> state(1U << sim->GetNumQubits());
     DataView<std::complex<double>, 1> view(state);
@@ -661,19 +657,13 @@ TEST_CASE("MatrixOperation test with 3-qubit", "[lightning]")
     CHECK(state[6].imag() == Approx(0.162104).epsilon(1e-5));
 }
 
-TEST_CASE("MatrixOperation test with 4-qubit", "[lightning]")
+TEMPLATE_TEST_CASE("MatrixOperation test with 4-qubit", "[GateSet]", LightningSimulator)
 {
-    std::unique_ptr<QuantumDevice> sim = CreateQuantumDevice();
+    std::unique_ptr<TestType> sim = std::make_unique<TestType>();
 
     // state-vector with #qubit = n
     constexpr size_t n = 4;
-    std::vector<QubitIdType> Qs;
-    Qs.reserve(n);
-
-    Qs[0] = sim->AllocateQubit();
-    Qs[1] = sim->AllocateQubit();
-    Qs[2] = sim->AllocateQubit();
-    Qs[3] = sim->AllocateQubit();
+    std::vector<QubitIdType> Qs = sim->AllocateQubits(n);
 
     sim->NamedOperation("Hadamard", {}, {Qs[0]}, false);
     sim->NamedOperation("CRX", {0.123}, {Qs[0], Qs[1]}, false);